initial update of tests
