+++ test/karma-cov.conf.js
@@ -3,12 +3,9 @@ module.exports = function (config) {
     browsers: ['Firefox'],
     frameworks: ['browserify', 'mocha'],
     reporters: ['mocha', 'coverage'],
-    preprocessors: { 'build/index.js': ['browserify'] },
-    browserify: {
-      debug: true,
-      transform: ['browserify-css', 'browserify-istanbul']
-    },
-    files: ['build/index.js'],
+    preprocessors: { 'build/*.js': ['browserify'] },
+    browserify: { debug: true, transform: ['browserify-css', 'browserify-istanbul'] },
+    files: ['build/*.js'],
     coverageReporter: {
       reporters : [
         { 'type': 'text' },
