+++ test/karma.conf.js
@@ -2,9 +2,9 @@ module.exports = function (config) {
   config.set({
     frameworks: ['browserify', 'mocha'],
     reporters: ['mocha'],
-    preprocessors: { 'build/index.js': ['browserify'] },
+    preprocessors: { 'build/*.js': ['browserify'] },
     browserify: { debug: true, transform: ['browserify-css'] },
-    files: ['build/index.js'],
+    files: ['build/*.js'],
     port: 9876,
     colors: true,
     singleRun: true,
