--- test/src/index.css
@@ -1,29 +1,0 @@
-/*-----------------------------------------------------------------------------
-| Copyright (c) 2014-2015, PhosphorJS Contributors
-|
-| Distributed under the terms of the BSD 3-Clause License.
-|
-| The full license is in the file LICENSE, distributed with this software.
-|----------------------------------------------------------------------------*/
-.offset-rect {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100px;
-  height: 100px;
-}
-
-
-.box-sizing {
-  border-top: solid 10px black;
-  border-left: solid 15px black;
-  padding: 15px 8px 9px 12px;
-}
-
-
-.size-limits {
-  min-width: 90px;
-  min-height: 95px;
-  max-width: 100px;
-  max-height: 105px;
-}
