+++ test/src/tsconfig.json
@@ -10,6 +10,9 @@
   "files": [
     "../../typings/expect.js/expect.js.d.ts",
     "../../typings/mocha/mocha.d.ts",
-    "index.ts"
+    "messages.ts",
+    "panel.ts",
+    "title.ts",
+    "widget.ts"
   ]
 }
